Move all functionality of displaying the list of Boxes to BoxFeed component.
