+++ src/app/(tabs)/index.tsx
@@ -1,58 +1,28 @@
 import React from 'react';
-import { FlatList, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { StyleSheet, Text, View } from 'react-native';
 
-import BoxComponent from '@/src/features/boxes/components/BoxComponent';
-import { Box, BoxesState } from '@/src/features/boxes/store/types';
 
-import { addBerry, removeBox } from '@/src/features/boxes/store/boxesSlice';
+import BoxFeed from '@/src/features/boxes/components/BoxFeed';
 import mockData from '@/src/mockdata/MockData.json';
 import { useRouter } from 'expo-router';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { useDispatch, useSelector } from 'react-redux';
 const routes = mockData[0].routes;
 const berries = mockData[1].berries;
 
 const IndexScreen = () => {
     const router = useRouter();
-    const dispatch = useDispatch();
-    const boxes: Box[] = useSelector(
-        (state: { boxes: BoxesState }) => state.boxes.items
-    );
 
-    const handleEditBerry = (boxId: string, newBerry: string) => {
-        dispatch(addBerry(boxId, newBerry))
-    };
-
-    const handleAddBox = () => {
+     const handleAddBox = () => {
         router.push('/(boxes)/createbox');
     };
-
-    const handleSelectBerry = () => {
-        console.log("Berry selected");
-    };
-
-    const handleSelectRoute = () => {
-        console.log("Route selected");
-    };
-
-    const handleRemoveBox = (boxId: string) => {
-        dispatch(removeBox(boxId));
-    }
 
     return (
         <SafeAreaView style={styles.safeArea}>
             <View>
-                <FlatList
-                    style={styles.list}
-                    contentContainerStyle={styles.listContent}
-                    data={boxes}
-                    keyExtractor={(item) => item.id}
-                    renderItem={({ item }) => <BoxComponent data={item} onEdit={(newBerry:string)=>handleEditBerry(item.id, newBerry)} onRemove={() => handleRemoveBox(item.id)} onSelectBerry={handleSelectBerry} onSelectRoute={handleSelectRoute}/>}
-                />
 
-                <TouchableOpacity activeOpacity={0.8} onPress={handleAddBox}>
-                    <Text style={styles.routeButton}>Add Box</Text>
-                </TouchableOpacity>
+                <BoxFeed />
+
+                <Text style={styles.routeButton} onPress={handleAddBox}>Add Box</Text>
             </View>
         </SafeAreaView>
     )
